<feat>: add stage quit to pause menu
- add stage quit to pause menu
- refactor contexts codes
