+++ src/contexts/stage/StageProvider.tsx
@@ -14,34 +14,11 @@ import { Quiz } from "@/types/quiz";
 import { DifficultyLevel } from "@/types/difficultyLevel";
 import { useMenuStore } from "@/store/useMenuStore";
 import { SCENE_IDS, useAppStore } from "@/store/useAppStore";
-import { getRandomQuizzesByDifficulty } from "@/utils/quizUtils";
 import { StageResultMenu } from "@/features/overlay/menu/components/StageResultMenu";
+import { StageState, UserAnswerResult } from "./StageTypes";
+import { Action, stageReducer } from "./stageReducer";
 
-// --- Types and Constants ---
-type STAGE_PHASES = [
-  "NONE",
-  "PREPARE",
-  "STAGE_START",
-  "ROUNDS_IN_PROGRESS",
-  "STAGE_RESULTS",
-  "STAGE_CONCLUDED",
-];
-
-type StagePhase = STAGE_PHASES[number];
-
-type UserAnswerResult = { wasCorrect: boolean; userAnswer: boolean } & Quiz;
-
-type StageState = {
-  phase: StagePhase;
-  difficultyLevel: DifficultyLevel | null;
-  quizzes: Quiz[];
-  correctCount: number;
-  currentRoundIndex: number;
-  lifePoints: number;
-  isPaused: boolean;
-  isVisibleTopHUD: boolean;
-};
-
+// --- Types and Constants  ---
 type StageContextType = {
   stageState: StageState;
   stageStateDispatch: Dispatch<Action>;
@@ -53,26 +30,10 @@ type StageContextType = {
     userAnswer: boolean;
   }) => void;
   setStageDifficulty: (level: DifficultyLevel) => void;
+  quitStage: () => void;
 };
 
-// --- Reducer and Actions ---
-
-type Action =
-  | { type: "RESET_TO_NONE" }
-  | { type: "SET_DIFFICULTY"; payload: DifficultyLevel }
-  | { type: "INIT_STAGE" }
-  | {
-      type: "PREPARE_STAGE";
-      payload: { allQuizzes: Quiz[]; difficulty: DifficultyLevel };
-    }
-  | { type: "START_STAGE" }
-  | {
-      type: "REPORT_ROUND_OUTCOME";
-      payload: { isCorrect: boolean; userAnswer: boolean };
-    }
-  | { type: "TOGGLE_PAUSE"; payload: boolean };
-
-const initialState: StageState = {
+export const initialState: StageState = {
   phase: "NONE",
   difficultyLevel: null,
   quizzes: [],
@@ -83,67 +44,9 @@ const initialState: StageState = {
   isVisibleTopHUD: false,
 };
 
-function stageReducer(stageState: StageState, action: Action): StageState {
-  switch (action.type) {
-    case "RESET_TO_NONE":
-      return { ...initialState };
-    case "SET_DIFFICULTY":
-      return { ...stageState, difficultyLevel: action.payload };
-    case "INIT_STAGE":
-      return { ...stageState, phase: "PREPARE" };
-    case "PREPARE_STAGE":
-      return {
-        ...initialState, // Reset stats for a new stage
-        difficultyLevel: action.payload.difficulty,
-        quizzes: getRandomQuizzesByDifficulty(
-          action.payload.allQuizzes,
-          action.payload.difficulty
-        ),
-        phase: "STAGE_START",
-        isVisibleTopHUD: true,
-      };
-    case "START_STAGE":
-      return { ...stageState, phase: "ROUNDS_IN_PROGRESS" };
-    case "REPORT_ROUND_OUTCOME": {
-      const { isCorrect } = action.payload;
-      const newCorrectCount = isCorrect
-        ? stageState.correctCount + 1
-        : stageState.correctCount;
-      const newLifePoints = isCorrect
-        ? stageState.lifePoints
-        : stageState.lifePoints - 1;
-
-      const isLastQuiz =
-        stageState.currentRoundIndex === stageState.quizzes.length - 1;
-      const isGameOver = newLifePoints === 0;
-
-      if (isLastQuiz || isGameOver) {
-        return {
-          ...stageState,
-          correctCount: newCorrectCount,
-          lifePoints: newLifePoints,
-          phase: "STAGE_RESULTS",
-          isVisibleTopHUD: false,
-        };
-      }
-      return {
-        ...stageState,
-        correctCount: newCorrectCount,
-        lifePoints: newLifePoints,
-        currentRoundIndex: stageState.currentRoundIndex + 1,
-      };
-    }
-    case "TOGGLE_PAUSE":
-      return { ...stageState, isPaused: action.payload };
-    default:
-      throw new Error("Unhandled action type in stageReducer");
-  }
-}
-
 // --- Context and Provider ---
 
 const StageContext = createContext<StageContextType | undefined>(undefined);
-
 interface StageProviderProps {
   children: ReactNode;
   totalQuizzes: Quiz[];
@@ -192,6 +95,10 @@ export const StageProvider = ({
     },
     [currentQuiz]
   );
+
+  const quitStage = () => {
+    stageStateDispatch({ type: "QUIT_STAGE" });
+  };
 
   // Effect for handling scene changes from the app store
   useEffect(() => {
@@ -244,7 +151,7 @@ export const StageProvider = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [stageState.phase]);
 
-  const contextValue = useMemo(
+  const contextValue = useMemo<StageContextType>(
     () => ({
       stageState,
       stageStateDispatch,
@@ -253,6 +160,7 @@ export const StageProvider = ({
       initStage,
       reportRoundOutcome,
       setStageDifficulty,
+      quitStage,
     }),
     [stageState, currentQuiz, reportRoundOutcome]
   );
