+++ src/features/game-canvas/components/sprites/BackgroundSprite.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { MAX_SIZE } from "@/constants/sizes";
-import { useRoundManager } from "@/contexts/RoundProvider";
+import { useRoundManager } from "@/contexts/round/RoundProvider";
 import {
   AnimatedSprite,
   AnimatedSpriteFrames,
