+++ src/features/overlay/action-bar/ActionBar.tsx
@@ -1,7 +1,7 @@
 // The refactored ActionBar component
 
 import { ActionButton } from "@/components/ui/button/ActionButton";
-import { useRoundManager } from "@/contexts/RoundProvider";
+import { useRoundManager } from "@/contexts/round/RoundProvider";
 import passButton from "@/assets/ui/button/pass_button.webp";
 import guardButton from "@/assets/ui/button/guard_button.webp";
